register metrics under internal route
