+++ .changeset/20260827234700-session-metrics-api.md
@@ -3,4 +3,4 @@
 '@truefoundry/trueforge': patch
 ---
 
-Add caller-scoped session metrics meters, charts, and chart-data under `/internal/sessions` via a server-owned `ISessionMetricsStore`.
+Add caller-scoped session metrics meters, charts, and chart-data under `/internal/metrics` via a server-owned `ISessionMetricsStore`.
